fix: hourly rate type

+++ database/migrations/2020_04_06_205934_add_hourly_rate_field_to_rooms_table.php
@@ -14,7 +14,7 @@ class AddHourlyRateFieldToRoomsTable extends Migration
     public function up()
     {
         Schema::table('rooms', function (Blueprint $table) {
-            $table->double('hourly_rate', 5, 2)->nullable();
+            $table->unsignedBigInteger('hourly_rate')->nullable();
         });
     }
